Trust the ClassSpecs
Summary:
They shouldn't be set if they're not true. I can't add assertions,
because there's no way of knowing why we know its safe to use the
Class*. For example we have method profiling that can decide to guard
on a particular Class*. It doesn't matter whether its persistent, or
if the context is a sub-type of it; we've guarded on that Class*, so
its guaranteed to be ok.

Reviewed By: paulbiss

Differential Revision: D3794749

fbshipit-source-id: c32cdf69ad3f82afa0bb6ba3b9401bc0392ea75a