Integration tests for IDE/disk updates interleaving
Summary: There was no coverage for this scenario. It also reveals a bug when after closing a file in IDE we don't revive disk contents until next file save. I asked Yue about it, and the motivation was that editor will always save a file to disk when closing, but that's not something that we should require it to do.

Reviewed By: alexchow

Differential Revision: D3757820

fbshipit-source-id: 942085110622609d316a3efccf1fe226f291cc0c