Split out send_push_message_to_client from send_response_to_client
Summary:
send_response_to_client is untypeable, because it's used to write arbitrary types to a file descriptor. By splitting sending push messages to separate function, we can give this function a type ("push", that I also create in this diff).

I also want this type to be more structured than "a random bundle of json", so I'm pushing conversion from errors to json futher into the client.

Reviewed By: jamesjwu

Differential Revision: D3766679

fbshipit-source-id: c9d22e6dd46ad07ebbd1e73cf416de370ea6aa3f