Improve IDE rechecking test
Summary: When debugging an issue, for a while I thought that I'm not cleaning up the list of files to recheck and end up rechecking every IDE file over and over. This turned out to not be the issue, but I'm sending out the test that I wrote to verify it anyway.

Reviewed By: jamesjwu

Differential Revision: D3769857

fbshipit-source-id: 0d28c72159f6ed8de111bd0267ca50f5b05f9cad