Fixed Datetime comparison with milliseconds involved
Summary:
I found this bug earlier in PHP and [fixed it there](https://github.com/php/php-src/pull/1653). After that I checked if it was also in HHVM. It was, so I also tried to fix it here :smile:

Depends on an updated version of [hhvm-third-party](https://github.com/hhvm/hhvm-third-party/pull/102).
Closes https://github.com/facebook/hhvm/pull/6888

Reviewed By: paulbiss

Differential Revision: D3034526

Pulled By: aorenste

fbshipit-source-id: 89a10c6af94002bc1992d73341ca5feeb7a2d1b2