Simplify [N]InstanceOfBitmask against constants
Summary:
Add a simplification role for InstanceOfBitmask and NInstanceOfBitmask where
both the object and class are constant. When generating an instanceof check, if
the class has instance bits allocated, we'll emit an InstanceOfBitmask op,
rather than a more generic ExtendsClass. However, if object is a constant, this
is actually a pessimization because we'll constant reduce ExtendsClass, but not
InstanceOfBitmask. Fix that.

Reviewed By: markw65

Differential Revision: D3802852

fbshipit-source-id: be802fffda4afe573aa1fe0a009825534071fcd8