Only clear stack frame state for inner-most frame
Summary:
During IR-gen, if we reach a block with an unprocessed predecessor, we clear our
frame state because we cannot account for what that predecessor will
do. However, if we're in an inlined function, we should only clear frame state
for the inner-most frame. We do this for most data, but not for the stack
state. This can cause problems, because when we return from the inlined
function, we've lost all the state for the caller's stack and the IR we generate
won't have the proper types. Fix this by only clearing stack state for the
inner-most frame, like the rest of the state.

Reviewed By: mxw

Differential Revision: D3813305

fbshipit-source-id: 7d120ff90bce6eca2e8787545bf771100f138a82