Don't overspecialize property gets and sets
Summary:
If we have an object property access, and guarding on the type of the
object would let us generate a direct load or store from the jit, we
would do it, even if we only needed to guarantee a base class, and we
were already guarding on that base class.

This adds a check to see if we're already guarding on the property's
owning class, and if so leaves it at that.

This still lets us over specialize if we weren't already guarding on a
base class. Fixing that is going to require allowing tracelet guards
on a subclass - where we currently only allow them on an exact
class. I think thats worthwhile - in many cases they're almost as
cheap as an exact class check - but its a much bigger change.

Reviewed By: swtaarrs

Differential Revision: D3817991

fbshipit-source-id: 92eb30770be790cb34e98e8dbc1c6afe4199bc22