Better type info for LdClsCtx
Summary:
I noticed that we weren't resolving get_called_class in non-static methods in
final classes, and its because LdCtx returns an Obj=foo|Cctx, and then LdClsCtx
doesn't know that the Cctx represents the same class as the Obj.

Also, in static methods in non-final classes, we weren't getting any class info
on the result of LdClsCtx because a Cctx can't represent subclass-of-foo.

Reviewed By: swtaarrs

Differential Revision: D3820154

fbshipit-source-id: 260b23e7d1039f160bdc0a368eef7fcec25d1c89