Move request stub free-list to stub-alloc.cpp
Summary:
The stub-alloc machinery is used exclusively from vasm and the request
handlers, so it doesn't belong directly in mcgen (logically the handlers sit
above mcgen in the stack, while vasm sits below). In a follow up to this stack
as I clarify the boundary between the request handlers and mcgen I hope to move
the locking out of freeRequestStub as I would like the CodeCache and associated
locks to reside solely within mcgen.

Reviewed By: mxw

Differential Revision: D3716248

fbshipit-source-id: 0ef650bd6c9d3a125a9ed6d1c84530a0404ec44d