Introduce tc module
Summary:
This diff splits out a new module which logically owns the code-cache, code and
metadata locks, and source-db. It is distinct from the mcgen module and divides
the JIT into three major components: runtime (service request handlers, etc),
mcgen, and tc. Runtime has access to the live VM state but does not directly
interact with the the irgen or vasm modules, and may not write into the tc
directly. The mcgen module is responsible for region selection, translation,
and profiling decisions, and the tc module is responsible for writing to the
TC.

Reviewed By: ottoni, swtaarrs

Differential Revision: D3756632

fbshipit-source-id: 40c7adc2bfc9da1d4754cb3e0466e1fbc0540c98