Convert FixupMap to namespace- remove fixupMap/sync from MCGenerator
Summary:
There was no reason that the FixupMap needed to be a class, nor was there any
reason that many of these functions and internal structures needed to be
exposed publicly. Create a FixupMap namespace that exposes a more limited
interface and move the sync functions into the fixup module from MCGenerator.

Reviewed By: markw65

Differential Revision: D3756640

fbshipit-source-id: d35d5902aa74968901e2eda364f87b38e35a0418