Kill MCGenerator \o/
Summary:
This is it. Kill MCGenerator, the mcg global, and the various includes thereof.
This diff introduces a flag it indicate mcgen has been initialized which
replaces the null checks we currently perform on mcg. It also required a lot of
include cleanup due to transitive includes via mc-generator.h.

Reviewed By: swtaarrs

Differential Revision: D3756663

fbshipit-source-id: 820a47b070339f217d8df2b6aea730d1f07c13f7