Upload load script logs to everstore in case of an exception/state not found
Summary:
Upload load script logs to everstore in case of an exception or state not found. This will make it easier for us to debug load script problems such as merge base not found or state problems.
Similarly, in the future if we aren't able to ssh into machines, this will also help us.

Reviewed By: jamesjwu

Differential Revision: D3825571

fbshipit-source-id: 2a82018d296dfd9ab445b3852d764aca632b2dbf