Improve lexing of XHP class names; now recognized in scope resolution expressions.
Summary:
We need to recognize ":foo::bar" as a legal scope-resolution expression, so we now parse terms using a special lexer that recognizes ":foo" as a token.

We still need to do more work here for cases like "class :foo extends :bar" and so on. Those will come in a later diff.

Reviewed By: jopamer

Differential Revision: D3779692

fbshipit-source-id: 5b09ad95ceb2173b9d93f1eb62dfe465fdff89fc