Allow function / const keywords in namespace use clause, part one
Summary:
The spec did not include the "group" form of the namespace use clause:

    use \Some\Prefix\ { C, const D, function E as F };

To prepare for parsing this form, I now allow the "function" and "const" to appear before the imported name in all cases, even though that is legal in only certain situations. We'll write an error checking pass that runs later to find the problem.

In the next diff we'll start adding gear to parse the group format. Note that this will require lexing namespace prefixes as tokens.

Reviewed By: andrewjkennedy

Differential Revision: D3788374

fbshipit-source-id: 5c703118bd82f184689df291777d7d31b62b2177