Fix overflowing integer multiplication for AArch64
AArch64 does not set flags on an integer-multiply and the common
way to detect an overflow on integer multiplication will only
provide EQ/NEQ flags.  To avoid taking a major penalty when
checking for overflows on integer multiplication, we make the
lowering of MulIntO architecture-specific and introduce a new
architecture-independent mulint (similarily to divint) vasm-insn.

The MulIntO node is lowered into the new 'mulint' vasm-instr, which
is a special kind of branching instruction (i.e. it takes a label
for the slow-path, in case an integer-overflow occurs).
'mulint' is marked to clobber the condition-code register.

The recommended way to detect an overflow on integer multiplication
on AArch64 is:
  mul	  calculate lower 64bit of the result
  smulh	  calculate the upper (signed/sign-extended) 64bits
  cmp	  compare if the upper 64bits are the sign-extension
	      of the lower 64bits (i.e. if all bits in the
 	      upper register are the same as the sign-bit in
	      the lower register.

To efficiently expand into such a sequence, the following vasm
instructions have been added:
  smulh	     upper 64bits of the 128bit result of a signed multiply
	       (could be named more nicely, but as only AArch64 is
	        using anyway...)
  cmpqsign   compare if one operand is the sign-extension of the
               other operand (i.e if all bits in one register are
               equivalent to the sign-bit of the other)
  mulint     integer-multiplication w/ overflow-checking

At the same time, the 'imul' vasm has been altered to _not_ set any
flags and _not_ perform any overflow-checking.  As it's only used
from the non-overflowing multiplication path in irlower-arith.c this
seems to be safe.

X-AffectedTests: quick/float-overflow.php
Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>