Make unwinder initialization more robust
Summary:
The recommended usage of hhvm.py is to source it from ~/.gdbinit.
However, gdb executes the contents of ~/.gdbinit before it has fully
started up and initialized all the Python data structures.  Since our
custom unwinder needs to access values in the inferior, and thus
requires the frame vector and symbol table to be set up, it can't be
safely run until after startup runs to completion.

This diff adds a command to manually initialize the unwinder, which
is also invoked whenever `walkstk` is called.  (Users of backtrace
will have to settle for the workflow of hitting `bt`, getting garbage,
and either running `unwind init` or `walkstk`.)

Reviewed By: markw65

Differential Revision: D3846453

fbshipit-source-id: d40475995fc15f6e5801506e9d412b1497175a25