Start allowing scope resolution in shape declarations and expressions
Summary:
The spec did not say that "foo::bar" expressions are legal as the left hand side of the arrow in a field initializer or field specifier of a shape.  That is:

    shape(foo::bar => int)  // legal type
    shape(foo::bar => 123)  // legal expression

I've modified the parser so that in the latter case we simply parse expressions on both sides of the =>.  We'll add a pass later to detect when the expression on the left is bogus.

For the "type" case I'll do the same thing but the fix will be a bit more involved, because right now the type parsing module has no dependency on the expression parsing module.  And we know how much fun it is to write recursive modules in OCaml.  That will come in the next diff.

Reviewed By: eshrews

Differential Revision: D3799972

fbshipit-source-id: ba3aaa9721c5901c80b3df75629cb498fc7ab478