Fix type parameter declarations
Summary:
Type parameter declarations were being parsed incorrectly; we were parsing the declared name as a type:

    class C<T>

The T there is a name; it cannot be a type:

    class C<?T>

for example is illegal, but we were allowing it.

Parsing this as a name token rather than a type changes a great many parse trees in tests; better now than later!

Also, the name parsed is permitted to be a keyword; we now allow oddities like

    class C<function>

to parse, as the previous parser did.

Reviewed By: eshrews, jopamer

Differential Revision: D3807473

fbshipit-source-id: 3d8e46d854709a87be74a687653cacf4b33e943f