Parse the "required" clause in an xhp attribute
Summary:
An xhp attribute declaration can end in required.

Interestingly, this is treated as one token by the type checker but two by hphp.  This causes a bug in the type checker:

    function required() : void {}
    ...
    required();

That should be legal, but the type checker parser rejects it.

The full fidelity parser treats it as two tokens.

Reviewed By: jopamer, eshrews

Differential Revision: D3814208

fbshipit-source-id: b9eb5d0601a51e7820997d63e2c4ceb5f429167e