Store the list of push diagnostic that need to be sent to client in ServerEnv
Summary:
Right now, push diagnostics is just sending the global list of all errors. I'll be changing it to be something more incremental, so that we can prioritize checking IDE files over their dependencies and disk updates, and send errors as they are available.

Those incremental errors, just like global errors are, will be computed in ServerTypeCheck.check. But I don't want ServerTypeCheck to know too much about clients, subscriptions and push messages - to avoid this, ServerTypeCheck is going to put the incremental errors in the environment, and they will be pushed to clients in ServerMain main loop.

Note: at the point of this diff env.errorl and env.diag_subscribe.errors are exactly the same, which looks redundant. But starting from the next diff I'll be diverging them to mean something slightly different.

Differential Revision: D3769202

fbshipit-source-id: 8e307e9c1b95064c7bc11984c4a59bc8d9d534b1