Change the internal representation of push diagnostics from list to map
Summary:
Error list is missing the ability to say "this file has no errors". This information is implicitly encoded in the fact that error list is global -  if you receive an error list, you know that all files in this list have errors, and all the other files don't.

Push diagnostics are (will be) different, more incremental. When you receive push diagnostics, you know that files in them have errors. But other files, that you received push diagnostics for before, might also still have errors.

In the latter model, we need to be able to express the concept of "this file, that had errors before (and you received diagnostics about before), no longer has errors". I'm doing it by making the diagnostics map from path to list of errors. Sending an empty list of errors for a path will mean that it has no errors anymore.

Differential Revision: D3769881

fbshipit-source-id: 611a5f5c5b9929683f8edf9a0c1a1e22e3bc73cc