Move filtering in ServerMain.recheck to Program.process_updates
Summary:
Just moving some stuff from one function to another to make the next diff
clearer. There is some filtering in both of those functions, and the only
reason it is split into two places is that this code used to be shared with
Flow, so there was common part, and project-specific part (e.g filtering ".php"
files versus ".js"). This is all ancient history and no longer applicable.

Reviewed By: dlreeves

Differential Revision: D3803849

fbshipit-source-id: 3ac898f26e6bd8b4ac652105ed54961240929d92