Don't pessimize vasm-copy when HHIRGenerateAsserts is on
Summary:
The vasm-copy pass for rewriting lea{}'s expects the lea{} that sets
rvmfp() in a call sequence to be immediately followed by a callphp{}.
When EvalHHIRGenerateAsserts is on, we emit code in-between, which
basically prevents vasm-copy from rewriting any memory operands with
rvmfp() as the base, which accounts for a lot of our stack pointer
rewriting.

Fix this by not doing the bad thing.

Reviewed By: swtaarrs

Differential Revision: D3860840

fbshipit-source-id: 2bc8e6d45549f9863741ae6dd7bd9bb6d78a6140