Remove parens from inclusion directive grammar
Summary:
The spec indicated that the grammar of the require / require_once inclusion directives' operand was either an expression or an expression surrounded by parentheses.  This is silly; a parenthesized expression is *already* an expression.

I've submitted a diff to the spec to eliminate this oddity, and removed it from the parser. The directive takes an expression; if that happens to be parenthesized, great.

This is part one of a series of diffs that will gradually make the inclusion directives into expressions; this is legal in non-strict mode.

Reviewed By: eshrews

Differential Revision: D3828342

fbshipit-source-id: fc9120cabc93702db95c8e4f2baaeef6b578d4a2