Add a new class to support TOC
This is the first commit of many to use a special feature for ppc:
TOC.

TOC (table of contents) is a pointer stored in a special register
(r2) that points to a memory region that contains immediate values
to be loaded throughout the jit execution. We are currently using
hhvm's data-block as this memomry region.

In this first commit we are not yet emitting code that use indirect
load to load immediates (all our changes in one commit would make
it big). For now, we just implemented the singleton class that
manages what is stored in TOC, and the "limmediate" function that
decides wether it is good or not to use TOC to load a specific
immediate value.