PageletServer: do not flush when output buffer closed
Summary:
Do not break internal unspoken invariants by modifying the state of already
closed output buffer. This is a quick fix to stop the crash, but perhaps we
should rethink how output buffers work in general. It feels wrong that we
silently ignore any attempt to echo/flush to a closed output buffer, ideally
that would throw a PHP exception. Also, there should be a generic mechanism to
check whether the output buffer is still open (not specific to the transport
type).

Reviewed By: bmaurer

Differential Revision: D3858598

fbshipit-source-id: a1c8cd8717c32c2b31178c9174f7cfa6787535ae