Split inclusion directive into directive, expression
Summary:
In non-strict mode we allow an inclusion directive to be used in a context where an expression is expected, except of course with no semicolon.

It is therefore easier to parse the directive as an inclusion expression followed by a semi.

We now do so. In an upcoming diff I will move the parsing logic to the expression parser, and ensure that the correct operator precedence is used.

Reviewed By: eshrews

Differential Revision: D3829195

fbshipit-source-id: 896cb84d48fea49b8046314fbb761b592d400793