GVN should not replace values with ones with less specific types
Summary:
GVN will look through instructions marked pass-through and potentially replace
any usages of the dest with the src. However, some pass-through instructions,
even though they conceptually do not change the value, may refine type
information. So, if GVN replaces a usage of a dest of a pass-through instruction
with that instruction's source, we could actually lose refined type
information. This may manifest itself as a pessimization, but it can also cause
incorrect IR to be generated. For example, if the usage of a dest of a
CastCtxThis instruction is replaced by the CastCtxThis' source, you've replaced
a value representing a TObj, with one that is TObj|TCctx, which isn't
equivalent. Fix all of this by only performing the substitution if the new
value's type is not less refined than the original.

Reviewed By: paulbiss

Differential Revision: D3868335

fbshipit-source-id: c37ff05ee68ff88b499dfd49a15de42d369c432b