Replace CastCtxThis with AssertType
Summary:
CastCtxThis is used to refine a TCtx into a TObj. However, to come up
with the proper TObj to return, it simply looks at the instruction's associated
func, and obtains an object type from that, totally ignoring the source
type. This actually can throw away information. For example, if a method in a
child class calls a method on the parent, and that call is inlined, the TObj
component of the source TCtx is more refined than CastCtxThis will infer on its
own (The TObj will be the child class, while CastCtxThis will infer the
parent). What CastCtxThis should really do is intersect the source type with its
own inferred type. However, at this point, what its doing is not very different
than AssertType, so just use that instead.

Reviewed By: swtaarrs

Differential Revision: D3868831

fbshipit-source-id: 25021aae68211772c12570375a368e1dc6583704