Split disk_needs_parsing set from ServerEnv.failed_parsing
Summary:
As the comment there points out, ServerEnv.failed_parsing conflates two
intentions: it's used simultaneously as

a) "the list of files which are known to have parsing errors in last round of
checking"
b) and "the list of files that just changed on disk and need to be re-parsed".

ServerMain.Program.recheck takes the list of disk updates and merges it with
pre-existing failed_parsing before triggering a recheck.

This conflation is fine, because recheck treats both of those sets of files the
same (it rechecks them), and since each recheck processes all the updates
completely, the contents of b) are not needed afterwards, so we can safely just
store a) in failed_parsing.

In the model I'm working toward, I don't want every recheck to necessary
process all the updates - I want to prioritize some of them while delaying
others. Now we need a place to store b) "for later".

Because of this, in this diff I stop merging the updates into failed_parsing
and pass them in a separate env field. ServerTypecheck then uses a helper
function to decide what it actually needs to parse based on those two fields.

As a result of this refactoring, ServerMain.Program.recheck and
ServerMain.recheck end up both being two lines long, so I merge them together
into one function.

Reviewed By: jamesjwu

Differential Revision: D3803847

fbshipit-source-id: 3528fe4d6a9a4e5af75ad79cd4dda9a4ec91b669