Properly re-attaches the LightProcess signal action after exec.
While executing a shell command, the ShellExecContext object
registers the default signal handler for the SIGCHLD signal by
calling signal(SIGCHLD, SIG_DFL). The default handler is a
traditional signal handler with a function prototype that takes
one parameter. After either a failed or successful attempt to exec
the shell command, the ShellExecContext dtor attempts to re-attach
the originally signal action originally registered by LigthProcess
by calling signal(SIGCHLD, m_sig_handler). While is is pretty
typical of sample code, this interface only supports the
re-attaching of a simple handler and not a signal action. The key
difference is that a signal action accepts three parameters.

Previously, this could would re-attach the saved signal action as
if it were a signal handler. Then when a signal arrived, the
signal action would be called with only on parameter and then trash
in the other two parameter registers. This fix uses a new method in
LightProcess to re-attach the signal action properly.

Fixes test/slow/ext_process/lwp.php on AArch64, although this is
a serious security bug on all platforms. We confirmed that the
x864_64 port on Ubuntu receives gargage for the second parameter
of the signal action for LightProcess::SigChldHandler().