Add lexer for XHP category names
Summary:
An XHP category name is an XHP name preceeded by "%". However, it is not *two tokens*; there must be no whitespace or comments between the % and the name.

This means that the language lexes ambiguously.  "category %foo" is two tokens but the expression "123 %foo" is three.

We therefore add a special lexer rule that treats %-followed-by-letter as a category name token, and everything else as a normal token.

In the next diff we'll set up data structures for the category parser, and then in the following diff we'll use the lexer to parse category declarations.

Reviewed By: dabek

Differential Revision: D3850926

fbshipit-source-id: 918a195db60d6b1c12bd4a8464661200194f0cb4