Parse "var" properties
Summary: PHP allows "var" as a synonym for "public" on a property declaration. This is illegal in Hack but we parse it anyways, for two reasons. First, so that we can do a good job if called upon to parse PHP in the future, and second, to give a good error message saying what to do in Hack. (The error message has not yet been implemented.)

Reviewed By: dabek

Differential Revision: D3863805

fbshipit-source-id: 859135aa21ff88f5a900b4b40f774503e2f07288