Enforce that allowed_classes in unserialize must be an array or a boolean.
Summary: In https://bugs.php.net/bug.php?id=72785 one of the changes that was made was to validate that the allowed_classes whitelist in unserialize was either a boolean or an array. I split out that work into its own diff for ease of review. Note that I needed to change the exception handling slightly to be compatible with the upstream behavior (documented here via a test)

Reviewed By: markw65

Differential Revision: D3869384

fbshipit-source-id: d43d07e498d6a9b354f5f2063f917824259c1abc