Data structures for parsing instanceof expressions
Summary:
"instanceof" was previously parsed by the ff-parser as just another infix binary operator. But there are special rules for the right hand side that are different than other expressions -- it can take either a type or a variable on the right -- and so we need special handling.

This diff just sets up the data structures; the parser and test case updates will be in the next diff.

Reviewed By: dabek

Differential Revision: D3864364

fbshipit-source-id: ef15d6b1598819689fe501dabf54db60dbee6022