Fix bad error recovery in full fidelity parser
Summary:
An important invariant of the FF parser is that *every* character in the file *must* be represented somewhere in the parse tree. We require this invariant because we only track widths of nodes; in order to know what text a node corresponds to, we must have an accurate count of all the nodes that come before it.

There was a mistake in the error recovery code for parsing methods; if the parser expected a method body and couldn't find one, then it just put the next token in the file into the parse tree as the *entire function*, and not *as the body*. This means that the function header was being dropped out of the parse tree at this point, and everything gets misaligned as a result.

Reviewed By: dabek

Differential Revision: D3865598

fbshipit-source-id: 49c6fd09a5f5173c9dd485a88610b39c0696562d