Data structures for soft types
Summary:
Hack apparently supports "soft" type specifiers; this feature is nowhere mentioned in the spec, so I'll have to add this one to the spec at a later date.

A type specifier may be preceded by "@" which apparently "softens" the errors reported if the hint is wrong.

This diff just sets up the data structures; the parsing and tests are in the next diff.

Differential Revision: D3866051

fbshipit-source-id: 03bcde9eb46e81a13468fea1a998619a73f760c4