Support errors coming from lazy checks in Diagnostic_subscription
Summary: Before D3803819, whenever we updated the list of errors, it was by replacing it with a completely new one. After that diff, the updates can be more fine grained - we need to know what were the files rechecked that generated new errors and update only the parts of map related to those files.

Reviewed By: jopamer

Differential Revision: D3876163

fbshipit-source-id: 7f1430aadc61b49ebfc6d5d0d20b7811bf382ba3