Make "stop at parsing errors" a property of check kind, not file kind
Summary:
We don't want mid-typing parse error to trigger invalidation of large swaths of dependent declarations. To achieve this, for IDE files we don't update parse trees if they had errors (so there will be no change to declarations).

But this behavior is not a property of "is this file open in IDE", but a property of "how fast do we want this check to finish", i.e "is this lazy check". In fact, we might want to have this behavior for files on disk - for example when closing a file in editor, we want to reparse disk file in lazy mode.

This diff refactors stuff to reflect this idea. It also allows to remove a bunch of copypasta from parsing_service, since the divergence between the two behaviors is now contained to a small function in ServerTypeCheck.

Reviewed By: jopamer, jamesjwu

Differential Revision: D3854040

fbshipit-source-id: 28392141feef8b3a1bfc9a80b1f8d85c9921f779