Remove "symbols_cache"
Summary:
This was a "proof of concept" on how can we speed up IDE operations thanks to
synced file contents. It's not deployed, and if we will want it, I think we
need to re-think it and do it more principled way than a bunch of "if cached
then ... else" sprinkled in multiple places. To make it maintainable it needs
to be lower in the stack and cached more opaquely to the tools using it.

As evidenced by changed in integration test, it wasn't even working correctly
in all cases.

Reviewed By: jopamer

Differential Revision: D3871871

fbshipit-source-id: 2fd56c0ec663eb936737acfb8ffcd5294f34102a