Support "close" file sync operation more correctly
Summary:
This operation doesn't really work - we don't reload disk contents after closing the file. The explanation I heard for it from yueli was that "editor will save the file to disk before closing it, so the disk recheck will end up reloading it".

Of course it is possible to close the file in editor without saving it, and we should support it.

Reviewed By: jopamer

Differential Revision: D3875870

fbshipit-source-id: 28bde60335bf4cccef077ecd483c8ccac59e9575