Eliminate unneeded helper methods
Summary:
I wish to make the API for the parse trees generated from a schema, which will require a number of cleanup steps.

One of them is to eliminate these helper methods; I thought they would be useful, but really just accessing the property is good enough.

This diff is just changing accesses to the helper methods into accesses to the fields, and deleting the helper methods.

Reviewed By: dabek

Differential Revision: D3879385

fbshipit-source-id: c21ece9e4c81ae5b2d8df330ca00ddb766b37ec4