Suppress "Unix_error(Unix.EPIPE, "write", _)" spam from logs
Summary: This exception used to be "Sys_error("Broken pipe")", but D3666204 changed the kind of exception thrown without updating the catch. I know that it might change back (alexchow has D3798572 out for that), but for now I don't think we have a firm decision on that, so I'm encapsulating it in a more local try-catch that should be less brittle in refactorings.

Reviewed By: jopamer

Differential Revision: D3880557

fbshipit-source-id: 7f629208578ed7fc3a9809e0a701369f3e48137e