Simplify how the unknown_objects list is processed
Summary:
After tracing the heap, we process the the set of unknown_objects_
by marking and scanning them eagerly. Then we resume tracing the heap.

It turns out the eager-scan part here is not necessary, similar to ordinary
roots. It suffices to enqueue the pointers, and let them be scanned by
process_worklist(), which will conservative-scan them.

Reviewed By: ricklavoie

Differential Revision: D3793446

fbshipit-source-id: aefe495b0dc252a28769025fe969ccaee779df83