Make data structures with one child consistent
Summary:
All "normal" parse tree nodes (that is, other than tokens, missing, error, and list) have one or more children, which are themselves parse tree nodes.  The "singleton" nodes -- things like variables, literals, etc -- have an "anonymous" child, and the nodes with two or more children have a record type that gives names to the children.

This makes it unnecessarily harder to treat parse tree nodes uniformly as things that have named children.  I've fixed this wart by introducing record types for the singletons.  This will make it easier to generate an API that represents these things in other languages; eventually all this boilerplate will be generated from a schema.

Reviewed By: dabek

Differential Revision: D3880206

fbshipit-source-id: 6e066614f6dddeec55c38fecd30ee0c1e71d8824